backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2

# LangChain core libs for RAG
langchain==0.2.16
langchain-community==0.2.11
langchain-core==0.2.38
langchain-text-splitters==0.2.2
# LangChain + Google Gemini integration
langchain-google-genai==1.0.7

# Google Generative AI (Gemini)
google-generativeai==0.8.2

# PDF processing
pypdf==4.3.1
pdf2image==1.17.0
Pillow==10.4.0

# Vector database
chromadb==0.5.5

# Embeddings
sentence-transformers==3.0.1

# Environment management
python-dotenv==1.0.1

# CORS provided via Starlette (dependency of FastAPI)
# Additional utilities
requests==2.32.3
aiofiles==24.1.0
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
requests-mock==1.12.1


